+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - [11681](https://github.com/vegaprotocol/vega/issues/11681) - Account for conflicts inserting funding payment records.
 - [11684](https://github.com/vegaprotocol/vega/issues/11684) - Better error when `Arbitrum` bridge details are missing from validator configuration.
 - [11696](https://github.com/vegaprotocol/vega/issues/11696) - Add binding for estimate fees API.
+- [11699](https://github.com/vegaprotocol/vega/issues/11699) - Update factors of programs when they are updated.
 
 
 ## 0.78.2